development/web/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
jinja2
openai
google-generativeai
pytest
httpx
python-multipart
pytest-asyncio
qdrant-client
